print_xfig.cpp: include cmath and explicit cast from std::floor (double) to int

+++ src-gui/p4/print_xfig.cpp
@@ -22,6 +22,8 @@
 #include <QFile>
 #include <QTextStream>
 
+#include <cmath>
+
 #include "P4InputVF.hpp"
 #include "P4ParentStudy.hpp"
 #include "custom.hpp"
@@ -394,10 +396,10 @@ static void xfig_print_line(double _x0, double _y0, double _x1, double _y1,
     if (sXFigBlackWhitePrint)
         color = printColorTable(bgColours::CFOREGROUND);
 
-    int x0{floor(_x0)};
-    int y0{floor(_y0)};
-    int x1{floor(_x1)};
-    int y1{floor(_y1)};
+    int x0{static_cast<int>(std::floor(_x0))};
+    int y0{static_cast<int>(std::floor(_y0))};
+    int x1{static_cast<int>(std::floor(_x1))};
+    int y1{static_cast<int>(std::floor(_y1))};
     if (x0 == x1 && y0 == y1)
         return;
 
@@ -627,161 +629,213 @@ static void xfig_print_point(double _x0, double _y0, int color)
 
 static void xfig_print_saddle(double x, double y)
 {
-    xfig_print_box(floor(x), floor(y), printColorTable(CSADDLE));
+    xfig_print_box(static_cast<int>(std::floor(x)),
+                   static_cast<int>(std::floor(y)), printColorTable(CSADDLE));
 }
 
 static void xfig_print_stablenode(double x, double y)
 {
-    xfig_print_box(floor(x), floor(y), printColorTable(CNODE_S));
+    xfig_print_box(static_cast<int>(std::floor(x)),
+                   static_cast<int>(std::floor(y)), printColorTable(CNODE_S));
 }
 static void xfig_print_unstablenode(double x, double y)
 {
-    xfig_print_box(floor(x), floor(y), printColorTable(CNODE_U));
+    xfig_print_box(static_cast<int>(std::floor(x)),
+                   static_cast<int>(std::floor(y)), printColorTable(CNODE_U));
 }
 
 static void xfig_print_stableweakfocus(double x, double y)
 {
-    xfig_print_diamond(floor(x), floor(y), printColorTable(CWEAK_FOCUS_S));
+    xfig_print_diamond(static_cast<int>(std::floor(x)),
+                       static_cast<int>(std::floor(y)),
+                       printColorTable(CWEAK_FOCUS_S));
 }
 
 static void xfig_print_unstableweakfocus(double x, double y)
 {
-    xfig_print_diamond(floor(x), floor(y), printColorTable(CWEAK_FOCUS_U));
+    xfig_print_diamond(static_cast<int>(std::floor(x)),
+                       static_cast<int>(std::floor(y)),
+                       printColorTable(CWEAK_FOCUS_U));
 }
 
 static void xfig_print_weakfocus(double x, double y)
 {
-    xfig_print_diamond(floor(x), floor(y), printColorTable(CWEAK_FOCUS));
+    xfig_print_diamond(static_cast<int>(std::floor(x)),
+                       static_cast<int>(std::floor(y)),
+                       printColorTable(CWEAK_FOCUS));
 }
 
 static void xfig_print_center(double x, double y)
 {
-    xfig_print_diamond(floor(x), floor(y), printColorTable(CCENTER));
+    xfig_print_diamond(static_cast<int>(std::floor(x)),
+                       static_cast<int>(std::floor(y)),
+                       printColorTable(CCENTER));
 }
 
 static void xfig_print_stablestrongfocus(double x, double y)
 {
-    xfig_print_diamond(floor(x), floor(y), printColorTable(CSTRONG_FOCUS_S));
+    xfig_print_diamond(static_cast<int>(std::floor(x)),
+                       static_cast<int>(std::floor(y)),
+                       printColorTable(CSTRONG_FOCUS_S));
 }
 
 static void xfig_print_unstablestrongfocus(double x, double y)
 {
-    xfig_print_diamond(floor(x), floor(y), printColorTable(CSTRONG_FOCUS_U));
+    xfig_print_diamond(static_cast<int>(std::floor(x)),
+                       static_cast<int>(std::floor(y)),
+                       printColorTable(CSTRONG_FOCUS_U));
 }
 
 static void xfig_print_sesaddle(double x, double y)
 {
-    xfig_print_triangle(floor(x), floor(y), printColorTable(CSADDLE));
+    xfig_print_triangle(static_cast<int>(std::floor(x)),
+                        static_cast<int>(std::floor(y)),
+                        printColorTable(CSADDLE));
 }
 
 static void xfig_print_sesaddlenode(double x, double y)
 {
-    xfig_print_triangle(floor(x), floor(y), printColorTable(CSADDLE_NODE));
+    xfig_print_triangle(static_cast<int>(std::floor(x)),
+                        static_cast<int>(std::floor(y)),
+                        printColorTable(CSADDLE_NODE));
 }
 
 static void xfig_print_sestablenode(double x, double y)
 {
-    xfig_print_triangle(floor(x), floor(y), printColorTable(CNODE_S));
+    xfig_print_triangle(static_cast<int>(std::floor(x)),
+                        static_cast<int>(std::floor(y)),
+                        printColorTable(CNODE_S));
 }
 
 static void xfig_print_seunstablenode(double x, double y)
 {
-    xfig_print_triangle(floor(x), floor(y), printColorTable(CNODE_U));
+    xfig_print_triangle(static_cast<int>(std::floor(x)),
+                        static_cast<int>(std::floor(y)),
+                        printColorTable(CNODE_U));
 }
 
 static void xfig_print_degen(double x, double y)
 {
-    xfig_print_cross(floor(x), floor(y), printColorTable(CDEGEN));
+    xfig_print_cross(static_cast<int>(std::floor(x)),
+                     static_cast<int>(std::floor(y)), printColorTable(CDEGEN));
 }
 
 static void xfig_print_virtualsaddle(double x, double y)
 {
     if (gVFResults.plotVirtualSingularities_)
-        xfig_print_box2(floor(x), floor(y), printColorTable(CSADDLE));
+        xfig_print_box2(static_cast<int>(std::floor(x)),
+                        static_cast<int>(std::floor(y)),
+                        printColorTable(CSADDLE));
 }
 
 static void xfig_print_virtualstablenode(double x, double y)
 {
     if (gVFResults.plotVirtualSingularities_)
-        xfig_print_box2(floor(x), floor(y), printColorTable(CNODE_S));
+        xfig_print_box2(static_cast<int>(std::floor(x)),
+                        static_cast<int>(std::floor(y)),
+                        printColorTable(CNODE_S));
 }
 static void xfig_print_virtualunstablenode(double x, double y)
 {
     if (gVFResults.plotVirtualSingularities_)
-        xfig_print_box2(floor(x), floor(y), printColorTable(CNODE_U));
+        xfig_print_box2(static_cast<int>(std::floor(x)),
+                        static_cast<int>(std::floor(y)),
+                        printColorTable(CNODE_U));
 }
 
 static void xfig_print_virtualstableweakfocus(double x, double y)
 {
     if (gVFResults.plotVirtualSingularities_)
-        xfig_print_diamond2(floor(x), floor(y), printColorTable(CWEAK_FOCUS_S));
+        xfig_print_diamond2(static_cast<int>(std::floor(x)),
+                            static_cast<int>(std::floor(y)),
+                            printColorTable(CWEAK_FOCUS_S));
 }
 
 static void xfig_print_virtualunstableweakfocus(double x, double y)
 {
     if (gVFResults.plotVirtualSingularities_)
-        xfig_print_diamond2(floor(x), floor(y), printColorTable(CWEAK_FOCUS_U));
+        xfig_print_diamond2(static_cast<int>(std::floor(x)),
+                            static_cast<int>(std::floor(y)),
+                            printColorTable(CWEAK_FOCUS_U));
 }
 
 static void xfig_print_virtualweakfocus(double x, double y)
 {
     if (gVFResults.plotVirtualSingularities_)
-        xfig_print_diamond2(floor(x), floor(y), printColorTable(CWEAK_FOCUS));
+        xfig_print_diamond2(static_cast<int>(std::floor(x)),
+                            static_cast<int>(std::floor(y)),
+                            printColorTable(CWEAK_FOCUS));
 }
 
 static void xfig_print_virtualcenter(double x, double y)
 {
     if (gVFResults.plotVirtualSingularities_)
-        xfig_print_diamond2(floor(x), floor(y), printColorTable(CCENTER));
+        xfig_print_diamond2(static_cast<int>(std::floor(x)),
+                            static_cast<int>(std::floor(y)),
+                            printColorTable(CCENTER));
 }
 
 static void xfig_print_virtualstablestrongfocus(double x, double y)
 {
     if (gVFResults.plotVirtualSingularities_)
-        xfig_print_diamond2(floor(x), floor(y),
+        xfig_print_diamond2(static_cast<int>(std::floor(x)),
+                            static_cast<int>(std::floor(y)),
                             printColorTable(CSTRONG_FOCUS_S));
 }
 
 static void xfig_print_virtualunstablestrongfocus(double x, double y)
 {
     if (gVFResults.plotVirtualSingularities_)
-        xfig_print_diamond2(floor(x), floor(y),
+        xfig_print_diamond2(static_cast<int>(std::floor(x)),
+                            static_cast<int>(std::floor(y)),
                             printColorTable(CSTRONG_FOCUS_U));
 }
 
 static void xfig_print_virtualsesaddle(double x, double y)
 {
     if (gVFResults.plotVirtualSingularities_)
-        xfig_print_triangle2(floor(x), floor(y), printColorTable(CSADDLE));
+        xfig_print_triangle2(static_cast<int>(std::floor(x)),
+                             static_cast<int>(std::floor(y)),
+                             printColorTable(CSADDLE));
 }
 
 static void xfig_print_virtualsesaddlenode(double x, double y)
 {
     if (gVFResults.plotVirtualSingularities_)
-        xfig_print_triangle2(floor(x), floor(y), printColorTable(CSADDLE_NODE));
+        xfig_print_triangle2(static_cast<int>(std::floor(x)),
+                             static_cast<int>(std::floor(y)),
+                             printColorTable(CSADDLE_NODE));
 }
 
 static void xfig_print_virtualsestablenode(double x, double y)
 {
     if (gVFResults.plotVirtualSingularities_)
-        xfig_print_triangle2(floor(x), floor(y), printColorTable(CNODE_S));
+        xfig_print_triangle2(static_cast<int>(std::floor(x)),
+                             static_cast<int>(std::floor(y)),
+                             printColorTable(CNODE_S));
 }
 
 static void xfig_print_virtualseunstablenode(double x, double y)
 {
     if (gVFResults.plotVirtualSingularities_)
-        xfig_print_triangle2(floor(x), floor(y), printColorTable(CNODE_U));
+        xfig_print_triangle2(static_cast<int>(std::floor(x)),
+                             static_cast<int>(std::floor(y)),
+                             printColorTable(CNODE_U));
 }
 
 static void xfig_print_virtualdegen(double x, double y)
 {
     if (gVFResults.plotVirtualSingularities_)
-        xfig_print_cross2(floor(x), floor(y), printColorTable(CDEGEN));
+        xfig_print_cross2(static_cast<int>(std::floor(x)),
+                          static_cast<int>(std::floor(y)),
+                          printColorTable(CDEGEN));
 }
 
 void xfig_print_coinciding(double x, double y)
 {
-    xfig_print_doublecross(floor(x), floor(y), printColorTable(CDEGEN));
+    xfig_print_doublecross(static_cast<int>(std::floor(x)),
+                           static_cast<int>(std::floor(y)),
+                           printColorTable(CDEGEN));
 }
 
 // ---------------------------------------------------------------------------------------
